# driver_service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Driver Service API
  version: "1.0"
  description: API for managing driver profiles

paths:
  /api/v1/drivers:
    post:
      summary: Create a new driver profile
      description: Creates a new driver profile and returns the created profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                firstName:
                  type: string
                phone:
                  type: string
                carNumber:
                  type: string
                fareType:
                  type: string
                status:
                  type: string
                rating:
                  type: integer
      responses:
        '201':
          description: Driver profile successfully created
          content:
            application/json:
              schema:
                type: object

    put:
      summary: Update an existing driver profile
      description: Updates an existing driver profile and returns the updated profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                firstName:
                  type: string
                phone:
                  type: string
                carNumber:
                  type: string
                fareType:
                  type: string
                status:
                  type: string
                rating:
                  type: integer
      responses:
        '200':
          description: Driver profile successfully updated
          content:
            application/json:
              schema:
                type: object

    get:
      summary: Get all drivers
      description: Retrieves a paginated list of drivers with sorting options.
      parameters:
        - name: phone
          in: query
          description: Phone number of the driver
          required: false
          schema:
            type: string
        - name: carNumber
          in: query
          description: Car number of the driver
          required: false
          schema:
            type: string
        - name: fare
          in: query
          description: Fare type to filter drivers
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Status to filter drivers
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          required: true
          schema:
            type: integer
        - name: size
          in: query
          description: Page size
          required: true
          schema:
            type: integer
        - name: orderBy
          in: query
          description: Field to sort by
          required: true
          schema:
            type: string
        - name: direction
          in: query
          description: Sort direction (ASC or DESC)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Driver not found

  # Получение водителя по ID
  /api/v1/drivers/{id}:
    get:
      summary: Get driver by ID
      description: Retrieves a driver by their unique identifier.
      parameters:
        - name: id
          in: path
          description: Unique identifier of the driver
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete a driver profile
      description: Deletes a driver profile by their unique identifier.
      parameters:
        - name: id
          in: path
          description: Unique identifier of the driver to delete
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Driver profile successfully deleted

  /api/v1/drivers/{id}/fare:
    get:
      summary: Get driver's fare type
      description: Retrieves the fare type of a driver by their unique identifier.
      parameters:
        - name: id
          in: path
          description: Unique identifier of the driver
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string

  /api/v1/drivers/{id}/status:
    get:
      summary: Get driver's status
      description: Retrieves the status of a driver by their unique identifier.
      parameters:
        - name: id
          in: path
          description: Unique identifier of the driver
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string

  /api/v1/drivers/{id}/rating:
    get:
      summary: Get driver's rating
      description: Retrieves the rating of a driver by their unique identifier.
      parameters:
        - name: id
          in: path
          description: Unique identifier of the driver
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer